+++ CHANGELOG.md
@@ -1,5 +1,9 @@
+## 1.1.0
+- fix Low Rendering Quality Issue #7
+- added `scaleRenderView` to `LayerBuffer` to compute the shader to a scaled resolution compared to the widget size.
+
 ## 1.0.1
-* fix GIF link in README
+- fix GIF link in README
 
 ## 1.0.0
-* Initial release.
+- Initial release.